… pom.xml
… src/main/java/com/researchspace/service/impl/RecordManagerImpl.java
… src/main/java/com/researchspace/service/impl/SearchManagerImpl.java
… src/main/java/com/researchspace/webapp/controller/ServiceLoggerAspct.java
… src/main/java/com/researchspace/webapp/controller/WorkspaceController.java
… src/test/java/com/researchspace/service/ExternalMessageHandlerImplTest.java
… src/test/java/com/researchspace/service/SearchManagerTest.java
… src/test/java/com/researchspace/slack/SlackPosterTest.java
… src/test/java/com/researchspace/webapp/controller/RecordManagerStub.java
… src/test/java/com/researchspace/webapp/controller/WorkspaceControllerPlainJunit.java
… src/test/java/com/researchspace/webapp/integrations/dmptool/DMPToolOAuthControllerMVCIT.java